Use real API in index.html
When I got around to implementing the real API isntead of the dummy data
I was using in my tests, I found that you can't just have the API return
the response to to the function that called it, because it happens
asynchronously. So you have to pass a callback function through, and
have the API function call that once the response has come back. If I'd
known that when I started I probably would have structured this a bit
differently. My HeadlinesController isn't doing very much for instance,
so it might have been better to have that and the relevent parts of the
APIInterface in the same object. I'm going to press on now though.

That's the first user story implemented though, with tests made
independent through lots of mocking and dependency injection, and a
feature test to check that everything is working together.

> As a busy politician
> I can see all of today's headlines in one place
> So I know what the big stories of the day are